Update e2e tests to use scaffolding containers
