--- test/fakeoidc/go.mod
@@ -1,7 +1,0 @@
-module github.com/sigstore/cosign/test/fakeoidc
-
-go 1.24.0
-
-require github.com/go-jose/go-jose/v4 v4.0.5
-
-require golang.org/x/crypto v0.45.0 // indirect
