+++ test/helpers.go
@@ -36,7 +36,6 @@ import (
 	"net/http/httptest"
 	"net/url"
 	"os"
-	"path"
 	"path/filepath"
 	"testing"
 	"time"
@@ -65,6 +64,7 @@ import (
 const (
 	rekorURL  = "http://127.0.0.1:3000"
 	fulcioURL = "http://127.0.0.1:5555"
+	tsaURL    = "http://127.0.0.1:3004"
 	certID    = "foo@bar.com"
 )
 
@@ -151,12 +151,13 @@ var verifyTSA = func(keyRef, imageRef string, checkClaims bool, annotations map[
 	return cmd.Exec(context.Background(), args)
 }
 
-var verifyKeylessTSA = func(imageRef string, tsaCertChain string, skipSCT bool, skipTlogVerify bool) error { //nolint: unused
+var verifyKeylessTSA = func(imageRef, tsaCertChain, certChain string, skipSCT, skipTlogVerify bool) error { //nolint: unused
 	cmd := cliverify.VerifyCommand{
 		CertVerifyOptions: options.CertVerifyOptions{
 			CertOidcIssuerRegexp: ".*",
 			CertIdentityRegexp:   ".*",
 		},
+		CertChain:        certChain,
 		RekorURL:         rekorURL,
 		HashAlgorithm:    crypto.SHA256,
 		TSACertChainPath: tsaCertChain,
@@ -510,13 +511,9 @@ func registryClientOpts(ctx context.Context) []remote.Option {
 
 // setLocalEnv sets SIGSTORE_CT_LOG_PUBLIC_KEY_FILE, SIGSTORE_ROOT_FILE, and SIGSTORE_REKOR_PUBLIC_KEY for the locally running sigstore deployment.
 func setLocalEnv(t *testing.T, dir string) error {
-	// fulcio repo is downloaded to the user's home directory by e2e_test.sh
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return fmt.Errorf("error getting home directory: %w", err)
-	}
-	t.Setenv(env.VariableSigstoreCTLogPublicKeyFile.String(), path.Join(home, "fulcio/config/ctfe/pubkey.pem"))
-	err = downloadAndSetEnv(t, fulcioURL+"/api/v1/rootCert", env.VariableSigstoreRootFile.String(), dir)
+	ctLogKey := os.Getenv("CT_LOG_KEY") //nolint: forbidigo
+	t.Setenv(env.VariableSigstoreCTLogPublicKeyFile.String(), ctLogKey)
+	err := downloadAndSetEnv(t, fulcioURL+"/api/v1/rootCert", env.VariableSigstoreRootFile.String(), dir)
 	if err != nil {
 		return fmt.Errorf("error setting %s env var: %w", env.VariableSigstoreRootFile.String(), err)
 	}
